feat: app #11 categoryContainer moved to component
